Restore unit states and module graphs from Buck cache
The new CLI option `--dep-units` may specify a JSON file containing all
build plans of the transitive dep closure, which can be used to restore
the unit states and module graph when recompiling with a freshly started
worker.

Also changes the interface of make mode to accept module names as an
alternative to source files, which causes `ModSummary` to be fetched
from the module graph.

Change-Id: Ica8b585222cc0cff88c0a377e8bfedfe97ee1ca9